refactor(celery): move Celery app instance to src/core/celery_app (#121)
Previously the Celery app was instantiated in src/auth/tasks, causing
every non-auth module (analytics, progress, school, reports, curriculum)
to carry a misleading auth dependency just to fire background tasks.

src/core/celery_app.py is now the canonical home for the Celery instance,
all task_routes, and the beat_schedule.  src/auth/tasks re-exports
celery_app and _run_async for backward compatibility; all nine caller files
updated to import directly from src.core.celery_app.  docker-compose.yml
worker commands updated from -A src.auth.tasks to -A src.core.celery_app.
